+++ src/sst/core/serialization/serialize_vector.h
@@ -1,8 +1,8 @@
-// Copyright 2009-2016 Sandia Corporation. Under the terms
+// Copyright 2009-2017 Sandia Corporation. Under the terms
 // of Contract DE-AC04-94AL85000 with Sandia Corporation, the U.S.
 // Government retains certain rights in this software.
 // 
-// Copyright (c) 2009-2016, Sandia Corporation
+// Copyright (c) 2009-2017, Sandia Corporation
 // All rights reserved.
 // 
 // This file is part of the SST software package. For license
@@ -45,7 +45,7 @@ class serialize<std::vector<T> > {
     }
     }
   
-    for (int i=0; i < v.size(); ++i){
+    for (size_t i=0; i < v.size(); ++i){
       serialize<T>()(v[i], ser);
     }
   }
