+++ src/sst/core/threadSyncSimpleSkip.cc
@@ -1,8 +1,8 @@
-// Copyright 2009-2016 Sandia Corporation. Under the terms
+// Copyright 2009-2017 Sandia Corporation. Under the terms
 // of Contract DE-AC04-94AL85000 with Sandia Corporation, the U.S.
 // Government retains certain rights in this software.
 // 
-// Copyright (c) 2009-2016, Sandia Corporation
+// Copyright (c) 2009-2017, Sandia Corporation
 // All rights reserved.
 // 
 // This file is part of the SST software package. For license
@@ -86,10 +86,10 @@ ThreadSyncSimpleSkip::before()
     // TraceFunction trace(CALL_INFO_LONG);
 
     // Empty all the queues and send events on the links
-    for ( int i = 0; i < queues.size(); i++ ) {
+    for ( size_t i = 0; i < queues.size(); i++ ) {
         ThreadSyncQueue* queue = queues[i];
         std::vector<Activity*>& vec = queue->getVector();
-        for ( int j = 0; j < vec.size(); j++ ) {
+        for ( size_t j = 0; j < vec.size(); j++ ) {
             Event* ev = static_cast<Event*>(vec[j]);
             auto link = link_map.find(ev->getLinkId());
             if (link == link_map.end()) {
@@ -142,7 +142,7 @@ ThreadSyncSimpleSkip::processLinkInitData()
     for ( int i = 0; i < num_threads; i++ ) {
         ThreadSyncQueue* queue = queues[i];
         std::vector<Activity*>& vec = queue->getVector();
-        for ( int j = 0; j < vec.size(); j++ ) {
+        for ( size_t j = 0; j < vec.size(); j++ ) {
             Event* ev = static_cast<Event*>(vec[j]);
             auto link = link_map.find(ev->getLinkId());
             if (link == link_map.end()) {
